+++ R/average_counts.R
@@ -49,7 +49,7 @@ if (getRversion() >= "2.15.1")
 #'    for this data file and tissue category.}
 #'  }
 #' @examples
-#' base_path = system.file("extdata", "TMA", package = "informr")
+#' base_path = system.file("extdata", "TMA", package = "phenoptr")
 #'
 #' # Count tumor cells near macrophages and tumor cells near CD8 separately,
 #' # in tumor and stroma tissue categories separately.