+++ R/read_cell_seg_data.R
@@ -31,7 +31,7 @@ list_cell_seg_files = function(path, ...) {
 #' \item Converts percent columns to numeric fractions
 #' \item Converts pixel distances to microns. The conversion factor may be
 #' specfied as a parameter or by setting
-#' \code{options(informr.pixels.per.micron)}.
+#' \code{options(phenoptr.pixels.per.micron)}.
 #' \item Optionally removes units from expression names
 #' \item If the file contains multiple sample names,
 #'       a \code{tag} column is created
@@ -52,7 +52,7 @@ list_cell_seg_files = function(path, ...) {
 #' @examples
 #' path = system.file("extdata", "TMA",
 #'                    "Core[1,5,6,1]_[21302,15107]_cell_seg_data.txt",
-#'                    package = "informr")
+#'                    package = "phenoptr")
 #' d = read_cell_seg_data(path)
 #'
 #' \dontrun{
@@ -62,7 +62,7 @@ list_cell_seg_files = function(path, ...) {
 #' }
 read_cell_seg_data <- function(
   path=NA,
-  pixels_per_micron=getOption('informr.pixels.per.micron'),
+  pixels_per_micron=getOption('phenoptr.pixels.per.micron'),
   remove_units=TRUE) {
   if (is.na(path)) {
     path <- file.choose()