+++ tests/testthat/test_read_cell_seg_data.R
@@ -3,7 +3,7 @@ library(testthat)
 
 path = system.file("extdata", "TMA",
    "Core[1,5,6,1]_[21302,15107]_cell_seg_data.txt",
-   package = "informr")
+   package = "phenoptr")
 
 expect_contains = function(container, items) {
   for (item in items)